# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "TraceNet API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/cases": {
   "get": {
    "summary": "List Cases",
    "operationId": "list_cases_cases_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/CaseOut"
         },
         "type": "array",
         "title": "Response List Cases Cases Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create Case",
    "operationId": "create_case_cases_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CaseCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CaseOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}": {
   "get": {
    "summary": "Get Case",
    "operationId": "get_case_cases__case_id__get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CaseOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/uploads": {
   "post": {
    "summary": "Upload Document",
    "operationId": "upload_document_cases__case_id__uploads_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_document_cases__case_id__uploads_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UploadOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/documents": {
   "post": {
    "summary": "Upload Document",
    "operationId": "upload_document_cases__case_id__documents_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_document_cases__case_id__documents_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UploadOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "summary": "List Documents",
    "operationId": "list_documents_cases__case_id__documents_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DocumentOut"
         },
         "title": "Response List Documents Cases  Case Id  Documents Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/jobs/{job_id}": {
   "get": {
    "summary": "Get Job",
    "operationId": "get_job_jobs__job_id__get",
    "parameters": [
     {
      "name": "job_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Job Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/stats": {
   "get": {
    "summary": "Case Stats",
    "description": "Dashboard headline numbers + cleaning summary (CaseStats shape).",
    "operationId": "case_stats_cases__case_id__stats_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CaseStatsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/documents/{document_id}/columns": {
   "get": {
    "summary": "Document Columns",
    "description": "Raw grid columns + samples for the guided column-mapping UI.",
    "operationId": "document_columns_documents__document_id__columns_get",
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Document Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DocumentColumnsOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/documents/{document_id}/template": {
   "post": {
    "summary": "Apply Template",
    "description": "Save the officer's column mapping as a bank template and re-parse.",
    "operationId": "apply_template_documents__document_id__template_post",
    "parameters": [
     {
      "name": "document_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Document Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ColumnTemplateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/templates": {
   "get": {
    "summary": "List Templates",
    "operationId": "list_templates_templates_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/BankTemplateOut"
         },
         "type": "array",
         "title": "Response List Templates Templates Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Save Template",
    "description": "Save an officer-defined column mapping for reuse on future statements.",
    "operationId": "save_template_templates_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BankTemplateIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/BankTemplateOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/transactions/{txn_id}/review": {
   "post": {
    "summary": "Review Transaction",
    "description": "Officer review: confirm / correct / exclude a row. Fully audit-logged.",
    "operationId": "review_transaction_transactions__txn_id__review_post",
    "parameters": [
     {
      "name": "txn_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Txn Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TransactionReview"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TransactionOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/clean": {
   "post": {
    "summary": "Clean Case",
    "description": "Run the cleaning pass: balance validation, duplicates, reversals.",
    "operationId": "clean_case_cases__case_id__clean_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Clean Case Cases  Case Id  Clean Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/analyze": {
   "post": {
    "summary": "Analyze Case",
    "description": "One-button analysis: cleaning \u2192 rules \u2192 ML \u2192 graph \u2192 round trips \u2192\ncorrelation \u2192 disposition. Returns the summary; artifacts via GETs below.",
    "operationId": "analyze_case_cases__case_id__analyze_post",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Analyze Case Cases  Case Id  Analyze Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/graph": {
   "get": {
    "summary": "Case Graph",
    "description": "Cytoscape.js elements: {nodes: [{data}], edges: [{data}]}.",
    "operationId": "case_graph_cases__case_id__graph_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/round-trips": {
   "get": {
    "summary": "Case Round Trips",
    "operationId": "case_round_trips_cases__case_id__round_trips_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/correlation": {
   "get": {
    "summary": "Case Correlation",
    "operationId": "case_correlation_cases__case_id__correlation_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/disposition": {
   "get": {
    "summary": "Case Disposition",
    "operationId": "case_disposition_cases__case_id__disposition_get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/cases/{case_id}/trail/{txn_id}": {
   "get": {
    "summary": "Money Trail",
    "description": "FIFO money trail for one credit (mentor requirement 5).",
    "operationId": "money_trail_cases__case_id__trail__txn_id__get",
    "parameters": [
     {
      "name": "case_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Case Id"
      }
     },
     {
      "name": "txn_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Txn Id"
      }
     },
     {
      "name": "stop_rule",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "tranche",
       "title": "Stop Rule"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Money Trail Cases  Case Id  Trail  Txn Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CaseCreate": {
    "properties": {
     "fir_number": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Fir Number"
     },
     "complainant": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Complainant"
     },
     "fraud_amount": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fraud Amount"
     },
     "complaint_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Complaint Date"
     }
    },
    "type": "object",
    "required": [
     "fir_number"
    ],
    "title": "CaseCreate"
   },
   "CaseOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "fir_number": {
      "type": "string",
      "title": "Fir Number"
     },
     "complainant": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Complainant"
     },
     "fraud_amount": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fraud Amount"
     },
     "complaint_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Complaint Date"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "fir_number",
     "complainant",
     "fraud_amount",
     "complaint_date",
     "created_at"
    ],
    "title": "CaseOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
  }
 }
}